+++ main_game/Assets/Scripts/Engineer/UpgradableTurret.cs
@@ -43,7 +43,7 @@ public class UpgradableTurret : UpgradableComponent
 	/// <returns>The efficiency as a value between 0 and 1.</returns>
 	public override float GetEfficiency()
 	{
-		int damage = MaxHealth - Health;
+		float damage = MaxHealth - Health;
 		return (MaxHealth - (damage / 2)) / MaxHealth;
 	}
